Added frontend handling of first app user case
If a user is the first one to use the application, they can't be
compared to any other users. The app was previously crashing
in this case, so lets make it clear to the user what's happening.

Also removed the example HomePage test now we have some
actual frontend tests.