Added basic analyzer which calculates typing speed
Using a simple analyzer now which calculates typing speed as the
total time taken (last event time - first event time) divided by
the number of characters typed (Keydown) events. Each time a user
completes a new test, the users new profile is the average of
the new test result and all previous results are taken to get
the users resulting profiles. In the future it may be worth
looking into using a median of results, as doing it this way
means the most recent test has a large impact.

From here we can setup a way of testing the algorithm to see if
future changes actually improve the algorithm.