Added testing script to check accuracy of results
Test data comes from Kaggle and is from 110 users typing
the string 'united states', around eight times each.
Using this script we can determine whether changes made
to the algorithm improve it's accuracy.

While I am unsure of the quality and suitability of this data
for this dummy application it should be ok, as gathering my
own data is not feasible.

I have included a script that transforms this data into a
format that the application uses. Half the data is used
as 'training data', while the other half is used as 'test data'.

For entry of test data, the top 3 most similar typist is calculated,
and we check if the actual user is in that top 3.