Fix tabbing potentially making test results inaccurate
As part of the analysis is based on speed, users tabbing
to write their name first and then coming back to write
the sentence means their speed will be calculated as
much slower than it actually is.

While there is probably a timeout between key presses that
after we can safely assume the user has stopped typing,
the most practical solution for now is to just remove
tab and enter characters from the dataset, as these are
not part of the sentences anyway.

Also disable form submission in the cases where the user
has not typed any characters, only spaces or tabs.